refactor: move imports to TYPE_CHECKING, lint fixes across modules
- backtest/report: pandas and RunContext to TYPE_CHECKING
- cli/commands: Callable to TYPE_CHECKING, inline RET504 returns
- config: sort __all__, add type: ignore for pydantic-settings
- features: extract quality adapter with SQLAlchemy connection wrapper
- features: prometheus metrics push in features_calc_short_service
- logging/utils: minor import cleanups

+++ src/cli/commands/label.py
@@ -158,8 +158,7 @@ async def _load_ohlcv(
         df.index = pd.to_datetime(df["timestamp"], unit="ms", utc=True)
         df.index.name = "timestamp"
         df = df.drop(columns=["timestamp"])
-        df = df.astype(float)
-        return df
+        return df.astype(float)
 
     except Exception as e:
         logger.error("Ошибка загрузки данных для %s %s: %s", symbol, timeframe, e)
